Modify and extract some submethods to shorten the main methods
Some of the methods are too long, so it would be appropriate to extract
segments of codes as submethods.
Previously, the methods stop adding new meals into user and main meal
list whenever an error in the .txt is encountered. Now, they have been
modified such that they simply print an error message, skip to the
next meal line in the .txt file and continue processing unless it is an
IOException.