Edit the compareLists()
It removes illegal meal found in the wishlist but not in recipes list.
This is abide by the rule that the meals present in the wishlist must
also be present in the recipes list and any meals remove from the
recipes list must also be removed from the wishlist.

A message is printed to inform the user of the illegal meal in wishlist.